pyproject.toml
[project]
name = "company-search-agent"
version = "0.1.0"
description = "Company research API using multiple data sources"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "google-api-python-client==2.108.0",
    "google-auth==2.23.4",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "langchain-openai>=0.3.11",
    "langchain-tavily>=0.2.0",
    "langgraph>=0.4.7",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.31.0",
    "ruff>=0.12.0",
    "tavily-python>=0.7.3",
    "uvicorn>=0.24.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
